:recycle: (plots) introduce a PlotManager
