+++ Commands/Record/Controller/Injected.cs
@@ -9,6 +9,7 @@ public partial class RecordController
 {
     public RecordFormatter Formatter { private get; set; } = new();
     public RecordManager Manager { private get; set; } = new();
+    public PlotManager PlotManager { private get; set; } = null!;
     public Random Random { private get; set; } = null!;
     public IKeyBasedCache<ulong, string> Cache { private get; set; } = null!;
 }
